refactor(components): refactor pagination
