+++ packages/components/pagination/src/usePagination.ts
@@ -1,13 +1,4 @@
 import { inject } from 'vue'
-import type { IPagination } from './pagination'
+import { elPaginationKey } from '@element-plus/tokens'
 
-export const usePagination = () => {
-  const pagination = inject<IPagination>('pagination', {})
-
-  return {
-    pagination,
-    pageCount: pagination.pageCount,
-    disabled: pagination.disabled,
-    currentPage: pagination.currentPage,
-  }
-}
+export const usePagination = () => inject(elPaginationKey, {})
