+++ packages/tokens/index.ts
@@ -1,3 +1,4 @@
 export * from './form'
 export * from './button'
 export * from './breadcrumb'
+export * from './pagination'
