+++ packages/utils/types.ts
@@ -49,3 +49,5 @@ export type TimeoutHandle = ReturnType<typeof global.setTimeout>
 export type ComponentSize = 'large' | 'medium' | 'small' | 'mini'
 
 export type StyleValue = string | CSSProperties | Array<StyleValue>
+
+export type Writeable<T> = { -readonly [P in keyof T]: T[P] }
